+++ src/User_Agent.php
@@ -56,6 +56,38 @@ class User_Agent {
     // -------------------------------------------------------------------------
     
     /**
+    * List operating systems and signatures
+    * 
+    * @var Array
+    */
+    protected static $operating_systems = array(
+        '/windows nt 10.0/i'    => 'Windows 10',
+        '/windows nt 6.2/i'     => 'Windows 8',
+        '/windows nt 6.1/i'     => 'Windows 7',
+        '/windows nt 6.0/i'     => 'Windows Vista',
+        '/windows nt 5.2/i'     => 'Windows Server 2003/XP x64',
+        '/windows nt 5.1/i'     => 'Windows XP',
+        '/windows xp/i'         => 'Windows XP',
+        '/windows nt 5.0/i'     => 'Windows 2000',
+        '/windows me/i'         => 'Windows ME',
+        '/win98/i'              => 'Windows 98',
+        '/win95/i'              => 'Windows 95',
+        '/win16/i'              => 'Windows 3.11',
+        '/macintosh|mac os x/i' => 'Mac OS X',
+        '/mac_powerpc/i'        => 'Mac OS 9',
+        '/linux/i'              => 'Linux',
+        '/ubuntu/i'             => 'Ubuntu',
+        '/iphone/i'             => 'iPhone',
+        '/ipod/i'               => 'iPod',
+        '/ipad/i'               => 'iPad',
+        '/android/i'            => 'Android',
+        '/blackberry/i'         => 'BlackBerry',
+        '/webos/i'              => 'Mobile',
+    );
+    
+    // -------------------------------------------------------------------------
+    
+    /**
     * List of devices and signatures
     * 
     * @var Array
@@ -188,6 +220,29 @@ class User_Agent {
             }
         }
         
+        return $name_when_no_match;
+    }
+    
+    // -------------------------------------------------------------------------
+    
+    /**
+    * Detects operating system according to user agent
+    * 
+    * @param String $user_agent
+    * @param String $name_when_no_match
+    * 
+    * @return String
+    */
+    public static function detect_operating_system($user_agent, $name_when_no_match='')
+    {
+        foreach (self::$operating_systems as $regex => $value)
+        {
+            if (preg_match($regex, $user_agent))
+            {
+                return $value;
+            }
+        }
+
         return $name_when_no_match;
     }